refactor(js-ts): migration init from jsx to tsx
