+++ package.json
@@ -41,7 +41,11 @@
     "devDependencies": {
         "gh-pages": "^2.0.1",
         "prettier": "^1.17.0",
-        "rimraf": "^3.0.2"
+        "rimraf": "^3.0.2",
+        "source-map-loader": "^1.1.3",
+        "ts-loader": "^8.0.12",
+        "typescript": "^4.1.3",
+        "webpack-cli": "^4.2.0"
     },
     "keywords": [
         "gatsby",
