--- src/pages/404.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-
-import Layout from '../components/layout/Layout';
-
-const IndexPage = () => (
-  <Layout fullMenu>
-    <article id="main">
-      <header>
-        <h2>Not Found</h2>
-        <p>Not a Valid URL</p>
-      </header>
-    </article>
-  </Layout>
-);
-
-export default IndexPage;
