+++ src/pages/Generic.tsx
